requirements.txt
alembic==1.8.0
anyio==3.6.1
asgiref==3.5.2
click==8.1.3
colorama==0.4.4
fastapi==0.78.0
FastAPI-SQLAlchemy==0.2.1
greenlet==1.1.2
h11==0.13.0
idna==3.3
importlib-metadata==4.11.4
importlib-resources==5.7.1
Mako==1.2.0
MarkupSafe==2.1.1
psycopg2==2.9.3
pydantic==1.9.1
python-dotenv==0.20.0
sniffio==1.2.0
SQLAlchemy==1.4.37
starlette==0.19.1
typing_extensions==4.2.0
uvicorn==0.17.6
zipp==3.8.0
